requirements.txt
pygame
paho-mqtt
python-etcd
